refactor(compute): fold clamp_table into _compute/runtime.py
build_clamp_active_table lived in _multi_compartment/clamp_table.py
but was imported from _compute/runtime.py at CellRuntimeState.from_cell
call time — a deferred import to break what was effectively a
layering inversion. Move both the dataclass and the builder into
runtime.py and delete the separate file. Tests absorbed into
runtime_test.py.
